Defence system, links, more admin features and minor improvements

+++ flaskapp.py
@@ -1,10 +1,13 @@
+__author__ = 'Alexey Maksimov (FroHenK)'
 import os
 import hashlib
 import json
 from datetime import datetime, timedelta
 
 from mysqlclient import curs, cnx, get_grades, get_subjects, get_homeworks, is_valid_admin, add_homework_db, \
-    create_day_element_db, get_day_db, get_day_db_id, get_homework_id, delete_day_db_id, delete_homework_id,get_all_day_db,get_all_homeworks
+    create_day_element_db, get_day_db, get_day_db_id, get_homework_id, delete_day_db_id, delete_homework_id, \
+    get_all_day_db, get_all_homeworks, add_subject_db, delete_subject_id
+
 from flask import Flask, request, session, flash, url_for, redirect, \
     render_template, abort, send_from_directory
 
@@ -13,17 +16,17 @@ app = Flask(__name__)
 app.config.from_pyfile('flaskapp.cfg')
 
 
-def template_imports():
+@app.context_processor
+def template_imports():  # put things into table for template rendering
     res = {}
 
     res["grades_array"] = get_grades()
-
     res["subjects_array"] = get_subjects()
     res["timedelta"] = timedelta
     return res
 
 
-@app.route('/set_grade/', methods=['POST'])
+@app.route('/set_grade/', methods=['POST'])  # set user's grade
 def set_grade():
     session['grade_selected_id'] = int(request.form['grade_id'])
     return redirect('/')
@@ -31,6 +34,8 @@ def set_grade():
 
 @app.route("/create_day_element/<string:usr_date>/", methods=['POST'])
 def create_day_element(usr_date):
+    if not session['is_admin']:
+        return not_enough_permissions()
     time_str = request.form['usr_time']
     subject_id = int(request.form['subject_id'])
     if len(time_str) == 0:
@@ -42,8 +47,10 @@ def create_day_element(usr_date):
 
 @app.before_request
 def b_request():
-    if session.get('grade_selected_id') is  None:
-        session['grade_selected_id'] = 105  # FIXME  bad code
+    if session.get('grade_selected_id') is None:  # if none grade is selected
+        session['grade_selected_id'] = 105  # FIXME bad code
+    if session.get('is_admin') is None:
+        session['is_admin'] = False
 
 
 @app.route("/get_day/<string:usr_date>/", methods=['GET'])
@@ -52,12 +59,13 @@ def get_homework(usr_date):
 
     day_db = get_day_db(session['grade_selected_id'], current_datetime)
     return render_template('schedule.html', day_elements_array=day_db, schedule_active=True,
-                           current_datetime=current_datetime,
-                           **template_imports())
+                           current_datetime=current_datetime)
 
 
 @app.route("/delete_day_element/<int:id>")
 def delete_day_elem(id):
+    if not session['is_admin']:
+        return not_enough_permissions()
     day_element = get_day_db_id(id)
     delete_day_db_id(id)
     return redirect('/get_day/' + day_element.date.strftime('%Y-%m-%d'))
@@ -65,6 +73,8 @@ def delete_day_elem(id):
 
 @app.route("/delete_homework/<int:id>")
 def delete_homework(id):
+    if not session['is_admin']:
+        return not_enough_permissions()
     day_element = get_homework_id(id)
     delete_homework_id(id)
     return redirect('/subjects/' + str(day_element.subject_id))
@@ -80,36 +90,61 @@ def admin_login():
 
 @app.route('/')
 def index():
-    # TODO make it
     current_datetime = datetime.now()
-
     return get_homework(current_datetime.strftime("%Y-%m-%d"))
 
 
 @app.route('/subjects/')
 def subjects_page():
-    return render_template('subjects.html', subjects_active=True, **template_imports())
+    return render_template('subjects.html', subjects_active=True)
 
 
 @app.route('/add_homework/<int:grade_id>/<int:subject_id>/', methods=["POST"])
 def on_add_homework(grade_id, subject_id):
+    if not session['is_admin']:
+        return not_enough_permissions()
+
     add_homework_db(grade_id, subject_id, str(request.form['data']))
 
     return redirect('/subjects/%s' % subject_id)
 
 
+@app.route('/add_subject/', methods=["POST"])
+def add_subject():
+    if not session['is_admin']:
+        return not_enough_permissions()
+
+    add_subject_db(str(request.form['title']))
+
+    return redirect('/subjects/')
+
+
+# not enough permissions
+def not_enough_permissions():
+    return render_template('not_enough_permissions.html')
+
+
+@app.route('/delete_subject/<int:subject_id>/')
+def remove_subject(subject_id):
+    if not session['is_admin']:
+        return not_enough_permissions()
+
+    delete_subject_id(subject_id)
+    return redirect(url_for('subjects_page'))
+
+
 @app.route('/subjects/<int:subject_id>/')
 def draw_subject(subject_id):
     return render_template('homeworks.html', subjects_active=True, subject_id=subject_id,
-                           homeworks_array=get_homeworks(session['grade_selected_id'], subject_id),
-                           **template_imports())
+                           homeworks_array=get_homeworks(session['grade_selected_id'], subject_id))
 
 
 @app.route('/project/')
 def about_project():
-    return render_template('project.html', project_active=True, **template_imports())
+    return render_template('project.html', project_active=True)
 
 
+'''
 @app.route('/admins')
 def get_admins():
     res = ""
@@ -118,6 +153,7 @@ def get_admins():
     for (id, username, password) in curs:
         res += "%s\n" % username.decode("utf8")
     return res
+'''
 
 
 @app.route('/<path:resource>')
@@ -125,14 +161,25 @@ def serve_static_resource(resource):
     return send_from_directory('static/', resource)
 
 
-def encode_b(obj):
+class JCalendar:  # how java's calendar looks in JSON
+    def __init__(self, year, month, dayOfMonth, hourOfDay, minute, second):
+        self.year = year
+        self.month = month - 1
+        self.dayOfMonth = dayOfMonth
+        self.hourOfDay = hourOfDay
+        self.minute = minute
+        self.second = second
+
+
+def encode_b(obj):  # JSON encode function
     if isinstance(obj, datetime):
-            serial = obj.isoformat()
-            return serial
+        serial = JCalendar(obj.year, obj.month, obj.day, obj.hour, obj.minute, obj.second)  # we need Java like calendar
+
+        return serial.__dict__
     return obj.__dict__
 
 
-@app.route("/get_json/")
+@app.route("/get_json/")  # API for Java written in JSON
 def json_api():
     json_data = {}
     json_data['grades'] = get_grades()
@@ -140,9 +187,9 @@ def json_api():
     json_data['homeworks'] = get_all_homeworks()
     json_data['day_elements'] = get_all_day_db()
 
-
     return json.dumps(json_data, default=encode_b)
 
 
 if __name__ == '__main__':
+    # add_subject_db('TestSubject')
     app.run()
